pyproject.toml
# pyproject.toml
[project]
name = "eve"
version = "0.1.0"
description = "Tool-calling creative assistant framework"
authors = [{ name = "Eden.art", email = "info@eden.art" }]
dependencies = [
    "anthropic>=0.49.0",
    "boto3>=1.35.64",
    "click>=8.1.7",
    "elevenlabs==2.10.0",
    "fastapi>=0.115.5",
    "gitpython>=3.1.43",
    "google-cloud-aiplatform>=1.115.0",
    "httpx>=0.27.2",
    "instructor==1.9.2",
    "jsonref>=1.1.0",
    "numpy>=2.1.3",
    "openai>=1.99.6,<1.100.0",
    "pillow>=11.0.0",
    "pydantic>=2.9.2",
    "pydub>=0.25.1",
    "pyjwt>=2.10.0",
    "pymongo>=4.10.0",
    "python-dotenv>=1.0.1",
    "python-magic>=0.4.27",
    "pyyaml>=6.0.2",
    "replicate>=1.0.7",
    "requests>=2.32.4",
    "runwayml>=3.11.0",
    "starlette>=0.41.3",
    "tenacity>=9.0.0",
    "tqdm>=4.67.0",
    "clerk-backend-api>=2.2.0,<3.0.0",
    "aiofiles>=24.1.0",
    "blurhash>=1.1.4",
    "py-cord>=2.4.1",
    "python-telegram-bot>=21.7",
    "requests-oauthlib>=1.3.1",
    "farcaster>=0.7.11",
    "ably>=2.0.7",
    "colorama>=0.4.6",
    "web3<7.6.1",
    "eth-account>=0.13.0",
    "psutil==6.0.0",
    "motor==3.3.1",
    "beanie==1.22.6",
    "apscheduler>=3.11.0",
    "modal>=1.0.1",
    "pytz>=2025.2",
    "fal-client==0.5.9",
    "torch==2.7.0",
    "safetensors==0.5.3",
    "moviepy>=2.2.1",
    "google-genai>=1.38.0",
    "langfuse==2.59.7",
    "litellm>=1.77.7",
    "tweepy>=4.15.0",
    "py-expression-eval>=0.3.14",
    "pyparsing>=3.2.3",
    "mcp>=1.13.0",
    "runway>=2.8.4",
    "transformers>=4.56.2",
    "pdfplumber>=0.11.0",
    "sentry-sdk>=2.41.0",
    "loguru>=0.7.3",
    "isodate>=0.6.1",
    "graphiti-core[falkordb,neptune]>=0.22.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = { find = { include = ["eve*"], exclude = [
    "tests*",
    "scripts*",
    "docker*"
] } }

[tool.ruff]
line-length = 88

[tool.ruff.lint]
ignore = ["E501", "F403", "E402", "F405"]
extend-select = ["I"]

[tool.rye]
dev-dependencies = [
    "uvicorn>=0.32.0",
    "ruff>=0.7.4",
    "pre-commit>=4.0.1",
    "pyright>=1.1.390",
    "pylance>=0.19.2",
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
]

[project.scripts]
eve = "eve.cli:cli"
check-imports = "scripts.check_imports:main"

[tool.pytest.ini_options]
markers = [
    "live: marks tests that require live API calls",
    "provider_anthropic: marks tests specific to Anthropic provider",
    "provider_gemini: marks tests specific to Gemini provider",
    "integration: marks tests that require integration setup",
    "asyncio: marks async tests (handled by pytest-asyncio plugin)"
]
